copy_block: Change type of array stride parameters to ptrdiff_t
ptrdiff_t is the correct type for array strides and similar.

+++ libavcodec/copy_block.h
@@ -19,11 +19,12 @@
 #ifndef AVCODEC_COPY_BLOCK_H
 #define AVCODEC_COPY_BLOCK_H
 
+#include <stddef.h>
 #include <stdint.h>
 
 #include "libavutil/intreadwrite.h"
 
-static inline void copy_block4(uint8_t *dst, const uint8_t *src, int dstStride, int srcStride, int h)
+static inline void copy_block4(uint8_t *dst, const uint8_t *src, ptrdiff_t dstStride, ptrdiff_t srcStride, int h)
 {
     int i;
     for (i = 0; i < h; i++) {
@@ -33,7 +34,7 @@ static inline void copy_block4(uint8_t *dst, const uint8_t *src, int dstStride, 
     }
 }
 
-static inline void copy_block8(uint8_t *dst, const uint8_t *src, int dstStride, int srcStride, int h)
+static inline void copy_block8(uint8_t *dst, const uint8_t *src, ptrdiff_t dstStride, ptrdiff_t srcStride, int h)
 {
     int i;
     for (i = 0; i < h; i++) {
@@ -43,7 +44,7 @@ static inline void copy_block8(uint8_t *dst, const uint8_t *src, int dstStride, 
     }
 }
 
-static inline void copy_block9(uint8_t *dst, const uint8_t *src, int dstStride, int srcStride, int h)
+static inline void copy_block9(uint8_t *dst, const uint8_t *src, ptrdiff_t dstStride, ptrdiff_t srcStride, int h)
 {
     int i;
     for (i = 0; i < h; i++) {
@@ -54,7 +55,7 @@ static inline void copy_block9(uint8_t *dst, const uint8_t *src, int dstStride, 
     }
 }
 
-static inline void copy_block16(uint8_t *dst, const uint8_t *src, int dstStride, int srcStride, int h)
+static inline void copy_block16(uint8_t *dst, const uint8_t *src, ptrdiff_t dstStride, ptrdiff_t srcStride, int h)
 {
     int i;
     for (i = 0; i < h; i++) {
@@ -64,7 +65,7 @@ static inline void copy_block16(uint8_t *dst, const uint8_t *src, int dstStride,
     }
 }
 
-static inline void copy_block17(uint8_t *dst, const uint8_t *src, int dstStride, int srcStride, int h)
+static inline void copy_block17(uint8_t *dst, const uint8_t *src, ptrdiff_t dstStride, ptrdiff_t srcStride, int h)
 {
     int i;
     for (i = 0; i < h; i++) {
